Refactored tool methods.
